Finished initial functionality
Adds make_accessible as both a class and instance method for ::Hash
Used as a class method, it defines method_missing for all Hash
instances, allowing dynamic getter and setter methods for the hash
elements. Used as an instance method, it has the same effect, but
restricted to the hash on which it was called.